+++ src/appliance-cooling.ts
@@ -8,6 +8,7 @@ import { HasChildLock } from './has-childlock';
 import { HasDoor } from './has-door';
 import { HasEvents } from './has-events';
 import { HasModes } from './has-modes';
+import { HasLight } from './has-light';
 import { Constructor } from './utils';
 
 // Add cooling appliance mode settings to an accessory
@@ -39,9 +40,10 @@ function HasCoolingModes<TBase extends Constructor<ApplianceBase>>(Base: TBase) 
 export class ApplianceFreezer extends
     HasChildLock(
     HasCoolingModes(
+    HasLight(
     HasEvents(
     HasDoor(
-    ApplianceGeneric)))) {
+    ApplianceGeneric))))) {
 
     constructor(...args: ConstructorParameters<typeof ApplianceGeneric>) {
         super(...args);
@@ -56,9 +58,10 @@ export class ApplianceFreezer extends
 export class ApplianceFridgeFreezer extends
     HasChildLock(
     HasCoolingModes(
+    HasLight(
     HasEvents(
     HasDoor(
-    ApplianceGeneric)))) {
+    ApplianceGeneric))))) {
 
     constructor(...args: ConstructorParameters<typeof ApplianceGeneric>) {
         super(...args);
@@ -74,9 +77,10 @@ export class ApplianceFridgeFreezer extends
 export class ApplianceRefrigerator extends
     HasChildLock(
     HasCoolingModes(
+    HasLight(
     HasEvents(
     HasDoor(
-    ApplianceGeneric)))) {
+    ApplianceGeneric))))) {
 
     constructor(...args: ConstructorParameters<typeof ApplianceGeneric>) {
         super(...args);